Stamp exporting, config based post-processors and IImage abstraction

+++ src/Core.Veldrid/Textures/VeldridMultiTexture.cs
@@ -12,7 +12,7 @@ namespace UAlbion.Core.Veldrid.Textures
     public class VeldridMultiTexture : MultiTexture, IVeldridTexture
     {
         public TextureType Type => TextureType.Texture2D;
-        public override uint FormatSize => Format.Size();
+        public override int FormatSize => Format.Size();
 
         // TODO: Cleanup
         public Texture CreateDeviceTexture(GraphicsDevice gd, ResourceFactory rf, TextureUsage usage)
@@ -24,7 +24,16 @@ namespace UAlbion.Core.Veldrid.Textures
                 RebuildLayers();
 
             var palette = PaletteManager.Palette.GetCompletePalette();
-            using var staging = rf.CreateTexture(new TextureDescription(Width, Height, Depth, MipLevels, ArrayLayers, Format.ToVeldrid(), TextureUsage.Staging, Type));
+            using var staging = rf.CreateTexture(new TextureDescription(
+                (uint)Width,
+                (uint)Height,
+                (uint)Depth,
+                (uint)MipLevels,
+                (uint)ArrayLayers,
+                Format.ToVeldrid(),
+                TextureUsage.Staging,
+                Type));
+
             staging.Name = "T_" + Name + "_Staging";
 
             Span<uint> toBuffer = stackalloc uint[(int)(Width * Height)];
@@ -45,9 +54,9 @@ namespace UAlbion.Core.Veldrid.Textures
                         fixed (uint* toBufferPtr = toBuffer)
                         {
                             gd.UpdateTexture(
-                                staging, (IntPtr)toBufferPtr, Width * Height * sizeof(uint),
+                                staging, (IntPtr)toBufferPtr, (uint)(Width * Height * sizeof(uint)),
                                 0, 0, 0,
-                                Width, Height, 1,
+                                (uint)Width, (uint)Height, 1,
                                 0, destinationLayer);
                         }
                     }
@@ -59,7 +68,16 @@ namespace UAlbion.Core.Veldrid.Textures
                 {
                 } //*/
 
-            var texture = rf.CreateTexture(new TextureDescription(Width, Height, Depth, MipLevels, ArrayLayers, Format.ToVeldrid(), usage, Type));
+            var texture = rf.CreateTexture(new TextureDescription(
+                (uint)Width,
+                (uint)Height,
+                (uint)Depth,
+                (uint)MipLevels,
+                (uint)ArrayLayers,
+                Format.ToVeldrid(),
+                usage,
+                Type));
+
             texture.Name = "T_" + Name;
 
             using (CommandList cl = rf.CreateCommandList())
